Revert "[GPU] Baseline InferRequest changes for dynamic shapes support (#12415)"
This reverts commit 7f665502b622eb07b85168f400f3fafceb096142.
