+++ src/plugins/intel_gpu/include/intel_gpu/plugin/compiled_model.hpp
@@ -32,9 +32,7 @@ public:
                                                                        InferenceEngine::OutputsDataMap networkOutputs) override;
     InferenceEngine::IInferRequestInternal::Ptr CreateInferRequestImpl(const std::vector<std::shared_ptr<const ov::Node>>& inputs,
                                                                        const std::vector<std::shared_ptr<const ov::Node>>& outputs) override;
-    template <class T>
-    InferenceEngine::IInferRequestInternal::Ptr GetInferRequestImpl(const std::vector<std::shared_ptr<const ov::Node>>& inputs,
-                                                                    const std::vector<std::shared_ptr<const ov::Node>>& outputs);
+
     InferenceEngine::Parameter GetMetric(const std::string &name) const override;
     InferenceEngine::Parameter GetConfig(const std::string &name) const override;
     std::shared_ptr<InferenceEngine::RemoteContext> GetContext() const override;
