+++ src/plugins/intel_gpu/include/intel_gpu/plugin/remote_context.hpp
@@ -63,13 +63,11 @@ public:
     std::shared_ptr<InferenceEngine::RemoteContext> getContext() const noexcept;
     InferenceEngine::LockedMemory<void> buffer() noexcept;
     InferenceEngine::LockedMemory<const void> cbuffer() const noexcept;
-    InferenceEngine::LockedMemory<void> rwmap() noexcept;
+    InferenceEngine::LockedMemory<void> rwmap()noexcept;
     InferenceEngine::LockedMemory<const void> rmap() const noexcept;
-    InferenceEngine::LockedMemory<void> wmap() noexcept;
+    InferenceEngine::LockedMemory<void> wmap()noexcept;
     const std::shared_ptr<InferenceEngine::IAllocator> &getAllocator() const noexcept;
     void *getHandle() const noexcept { return _handle; }
-
-    void reinterpret(cldnn::layout new_layout);
 
     bool is_allocated() const noexcept;
     bool is_locked() const noexcept;
