+++ src/plugins/intel_gpu/include/intel_gpu/runtime/layout.hpp
@@ -482,8 +482,6 @@ struct layout {
 
     void set_tensor(const tensor& size);
 
-    void set_partial_shape(const ov::PartialShape& size);
-
     // Returns true if other layout can be reinterpreted without need of reordering
     bool compatible(const layout& other) const;
