+++ src/plugins/intel_gpu/src/plugin/compiled_model.cpp
@@ -8,7 +8,6 @@
 #include "intel_gpu/plugin/infer_request.hpp"
 #include "intel_gpu/plugin/compiled_model.hpp"
 #include "intel_gpu/plugin/async_infer_request.hpp"
-#include "intel_gpu/plugin/async_infer_request_legacy.hpp"
 #include "openvino/runtime/intel_gpu/properties.hpp"
 
 #include <description_buffer.hpp>
@@ -64,32 +63,19 @@ CompiledModel::CompiledModel(InferenceEngine::CNNNetwork &network, std::shared_p
     }
 }
 
-template <class T>
-IInferRequestInternal::Ptr CompiledModel::GetInferRequestImpl(const std::vector<std::shared_ptr<const ov::Node>>& inputs,
-                                                              const std::vector<std::shared_ptr<const ov::Node>>& outputs) {
-    auto ptr = std::make_shared<T>(inputs, outputs, std::static_pointer_cast<CompiledModel>(shared_from_this()));
-    if (m_config.throughput_streams > 1)
-        ptr->EnableStreams();
-    if (m_config.useProfiling)
-        ptr->EnableProfiling();
-    if (m_graphs.front()->use_external_queue())
-        ptr->enable_external_queue();
-    ptr->SetGraph(m_graphs.front());
-
-    return ptr;
-}
-
 IInferRequestInternal::Ptr CompiledModel::CreateInferRequestImpl(InputsDataMap networkInputs,
                                                                  OutputsDataMap networkOutputs) {
     OV_ITT_SCOPED_TASK(itt::domains::intel_gpu_plugin, "CompiledModel::CreateInferRequestImpl");
-    auto ptr = std::make_shared<InferRequestLegacy>(networkInputs, networkOutputs,
-                                                    std::static_pointer_cast<CompiledModel>(shared_from_this()));
-    if (m_config.throughput_streams > 1)
+    auto ptr = std::make_shared<InferRequest>(networkInputs, networkOutputs,
+                                              std::static_pointer_cast<CompiledModel>(shared_from_this()));
+    if (m_config.throughput_streams > 1) {
         ptr->EnableStreams();
+    }
     if (m_config.useProfiling)
         ptr->EnableProfiling();
-    if (m_graphs.front()->use_external_queue())
+    if (m_graphs.front()->use_external_queue()) {
         ptr->enable_external_queue();
+    }
     ptr->SetGraph(m_graphs.front());
 
     return ptr;
@@ -98,10 +84,20 @@ IInferRequestInternal::Ptr CompiledModel::CreateInferRequestImpl(InputsDataMap n
 IInferRequestInternal::Ptr CompiledModel::CreateInferRequestImpl(const std::vector<std::shared_ptr<const ov::Node>>& inputs,
                                                                  const std::vector<std::shared_ptr<const ov::Node>>& outputs) {
     OV_ITT_SCOPED_TASK(itt::domains::intel_gpu_plugin, "CompiledModel::CreateInferRequestImpl");
-    if (m_graphs.front()->GetMaxDynamicBatchSize() > 1)
-        return GetInferRequestImpl<InferRequestLegacy>(inputs, outputs);
-    else
-        return GetInferRequestImpl<InferRequest>(inputs, outputs);
+    auto ptr = std::make_shared<InferRequest>(inputs, outputs,
+                                              std::static_pointer_cast<CompiledModel>(shared_from_this()));
+    if (m_config.throughput_streams > 1) {
+        ptr->EnableStreams();
+    }
+    if (m_config.useProfiling)
+        ptr->EnableProfiling();
+
+    if (m_graphs.front()->use_external_queue()) {
+        ptr->enable_external_queue();
+    }
+    ptr->SetGraph(m_graphs.front());
+
+    return ptr;
 }
 
 IInferRequestInternal::Ptr CompiledModel::CreateInferRequest() {
@@ -121,23 +117,12 @@ IInferRequestInternal::Ptr CompiledModel::CreateInferRequest() {
         }
     }
 
-    bool is_legacy = false;
     if (this->_plugin && _plugin->IsNewAPI()) {
         internalRequest = CreateInferRequestImpl(_parameters, _results);
-        if (std::dynamic_pointer_cast<InferRequestLegacy>(internalRequest))
-            is_legacy = true;
     }
-    if (!internalRequest) {
+    if (!internalRequest)
         internalRequest = CreateInferRequestImpl(_networkInputs, _networkOutputs);
-        is_legacy = true;
-    }
     internalRequest->setPointerToExecutableNetworkInternal(shared_from_this());
-    if (is_legacy) {
-        return std::make_shared<AsyncInferRequestLegacy>(std::static_pointer_cast<InferRequestLegacy>(internalRequest),
-                                                         m_taskExecutor,
-                                                         m_waitExecutor,
-                                                         _callbackExecutor);
-    }
     return std::make_shared<AsyncInferRequest>(std::static_pointer_cast<InferRequest>(internalRequest),
                                                m_taskExecutor,
                                                m_waitExecutor,
