+++ src/plugins/intel_gpu/src/plugin/remote_context.cpp
@@ -194,14 +194,6 @@ std::shared_ptr<InferenceEngine::RemoteContext> RemoteBlobImpl::getContext() con
     return m_context.lock();
 }
 
-void RemoteBlobImpl::reinterpret(cldnn::layout new_layout) {
-    OPENVINO_ASSERT(m_layout.bytes_count() >= new_layout.bytes_count(),
-                    "[GPU] Can't reinterpret blob to the size bigger than allocated memory buffer");
-    m_layout = new_layout;
-    auto engine = m_memObject->get_engine();
-    m_memObject = engine->reinterpret_buffer(*m_memObject, new_layout);
-}
-
 void RemoteBlobImpl::lock() const {
     if (!is_allocated()) {
         IE_THROW(NotAllocated) << "[GPU] Remote blob can't be locked as it's not allocated";
@@ -242,7 +234,7 @@ LockedMemory<const void> RemoteBlobImpl::cbuffer() const noexcept {
     }
 }
 
-LockedMemory<void> RemoteBlobImpl::rwmap() noexcept {
+LockedMemory<void> RemoteBlobImpl::rwmap()noexcept {
     try {
         lock();
         return LockedMemory<void>(reinterpret_cast<IAllocator *>(&m_allocator), _handle, 0);
@@ -260,7 +252,7 @@ LockedMemory<const void> RemoteBlobImpl::rmap() const noexcept {
     }
 }
 
-LockedMemory<void> RemoteBlobImpl::wmap() noexcept {
+LockedMemory<void> RemoteBlobImpl::wmap()noexcept {
     try {
         lock();
         return LockedMemory<void>(reinterpret_cast<IAllocator *>(&m_allocator), _handle, 0);
