+++ src/tests/functional/plugin/gpu/shared_tests_instances/behavior/ov_infer_request/infer_request_dynamic.cpp
@@ -11,32 +11,13 @@ using namespace ov::test::behavior;
 
 namespace {
 
-const std::vector<ov::AnyMap> configs = {
-    {}
-};
-
 const std::vector<ov::AnyMap> AutoConfigs = {
     {ov::device::priorities(CommonTestUtils::DEVICE_GPU, CommonTestUtils::DEVICE_CPU)},
     {}
 };
-
 const std::vector<ov::AnyMap> AutoNotSupportConfigs = {
     {ov::device::priorities(CommonTestUtils::DEVICE_GPU)}
 };
-
-std::shared_ptr<ngraph::Function> getFunction1() {
-    const std::vector<size_t> inputShape = {1, 4, 20, 20};
-    const ngraph::element::Type_t ngPrc = ngraph::element::Type_t::f32;
-
-    auto params = ngraph::builder::makeParams(ngPrc, {inputShape});
-    params.front()->set_friendly_name("Param_1");
-    params.front()->get_output_tensor(0).set_names({"input_tensor"});
-
-    auto relu = std::make_shared<ngraph::opset1::Relu>(params[0]);
-    relu->get_output_tensor(0).set_names({"relu"});
-
-    return std::make_shared<ngraph::Function>(relu, params, "SimpleActivation");
-}
 
 std::shared_ptr<ngraph::Function> getFunction2() {
     const std::vector<size_t> inputShape = {1, 4, 20, 20};
@@ -60,16 +41,6 @@ std::shared_ptr<ngraph::Function> getFunction2() {
 
     return std::make_shared<ngraph::Function>(concat, params, "SplitAddConcat");
 }
-
-INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_1, OVInferRequestDynamicTests,
-                        ::testing::Combine(
-                                ::testing::Values(getFunction1()),
-                                ::testing::Values(std::vector<std::pair<std::vector<size_t>, std::vector<size_t>>>{
-                                    {{1, 4, 20, 20}, {1, 4, 20, 20}},
-                                    {{2, 4, 20, 20}, {2, 4, 20, 20}}}),
-                                ::testing::Values(CommonTestUtils::DEVICE_GPU),
-                                ::testing::ValuesIn(configs)),
-                        OVInferRequestDynamicTests::getTestCaseName);
 
 INSTANTIATE_TEST_SUITE_P(smoke_Auto_BehaviorTests, OVInferRequestDynamicTests,
                         ::testing::Combine(
